refactored Admin Dash
