+++ client/components/navbar.js
@@ -9,7 +9,7 @@ const Navbar = ({handleClick, isLoggedIn, cart}) => (
     <ul>
       <li>
         <NavLink id="logoMainNav" to="/home">
-          <img src="./pictures/gslogopic.jpg" />
+          <img src="/pictures/gslogopic.jpg" />
         </NavLink>
         <p>UTONIUM</p>
       </li>
